.pub-cache/hosted/pub.dartlang.org/html-0.15.0/test/data/tree-construction/tests_innerHTML_1.dat

#data
<body><span>
#errors
#document-fragment
body
#document
| <span>

#data
<span><body>
#errors
#document-fragment
body
#document
| <span>

#data
<span><body>
#errors
#document-fragment
div
#document
| <span>

#data
<body><span>
#errors
#document-fragment
html
#document
| <head>
| <body>
|   <span>

#data
<frameset><span>
#errors
#document-fragment
body
#document
| <span>

#data
<span><frameset>
#errors
#document-fragment
body
#document
| <span>

#data
<span><frameset>
#errors
#document-fragment
div
#document
| <span>

#data
<frameset><span>
#errors
#document-fragment
html
#document
| <head>
| <frameset>

#data
<table><tr>
#errors
#document-fragment
table
#document
| <tbody>
|   <tr>

#data
</table><tr>
#errors
#document-fragment
table
#document
| <tbody>
|   <tr>

#data
<a>
#errors
#document-fragment
table
#document
| <a>

#data
<a>
#errors
#document-fragment
table
#document
| <a>

#data
<a><caption>a
#errors
#document-fragment
table
#document
| <a>
| <caption>
|   "a"

#data
<a><colgroup><col>
#errors
#document-fragment
table
#document
| <a>
| <colgroup>
|   <col>

#data
<a><tbody><tr>
#errors
#document-fragment
table
#document
| <a>
| <tbody>
|   <tr>

#data
<a><tfoot><tr>
#errors
#document-fragment
table
#document
| <a>
| <tfoot>
|   <tr>

#data
<a><thead><tr>
#errors
#document-fragment
table
#document
| <a>
| <thead>
|   <tr>

#data
<a><tr>
#errors
#document-fragment
table
#document
| <a>
| <tbody>
|   <tr>

#data
<a><th>
#errors
#document-fragment
table
#document
| <a>
| <tbody>
|   <tr>
|     <th>

#data
<a><td>
#errors
#document-fragment
table
#document
| <a>
| <tbody>
|   <tr>
|     <td>

#data
<table></table><tbody>
#errors
#document-fragment
caption
#document
| <table>

#data
</table><span>
#errors
#document-fragment
caption
#document
| <span>

#data
<span></table>
#errors
#document-fragment
caption
#document
| <span>

#data
</caption><span>
#errors
#document-fragment
caption
#document
| <span>

#data
<span></caption><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><caption><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><col><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><colgroup><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><html><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><tbody><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><td><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><tfoot><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><thead><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><th><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><tr><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span></table><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
</colgroup><col>
#errors
#document-fragment
colgroup
#document
| <col>

#data
<a><col>
#errors
#document-fragment
colgroup
#document
| <col>

#data
<caption><a>
#errors
#document-fragment
tbody
#document
| <a>

#data
<col><a>
#errors
#document-fragment
tbody
#document
| <a>

#data
<colgroup><a>
#errors
#document-fragment
tbody
#document
| <a>

#data
<tbody><a>
#errors
#document-fragment
tbody
#document
| <a>

#data
<tfoot><a>
#errors
#document-fragment
tbody
#document
| <a>

#data
<thead><a>
#errors
#document-fragment
tbody
#document
| <a>

#data
</table><a>
#errors
#document-fragment
tbody
#document
| <a>

#data
<a><tr>
#errors
#document-fragment
tbody
#document
| <a>
| <tr>

#data
<a><td>
#errors
#document-fragment
tbody
#document
| <a>
| <tr>
|   <td>

#data
<a><td>
#errors
#document-fragment
tbody
#document
| <a>
| <tr>
|   <td>

#data
<a><td>
#errors
#document-fragment
tbody
#document
| <a>
| <tr>
|   <td>

#data
<td><table><tbody><a><tr>
#errors
#document-fragment
tbody
#document
| <tr>
|   <td>
|     <a>
|     <table>
|       <tbody>
|         <tr>

#data
</tr><td>
#errors
#document-fragment
tr
#document
| <td>

#data
<td><table><a><tr></tr><tr>
#errors
#document-fragment
tr
#document
| <td>
|   <a>
|   <table>
|     <tbody>
|       <tr>
|       <tr>

#data
<caption><td>
#errors
#document-fragment
tr
#document
| <td>

#data
<col><td>
#errors
#document-fragment
tr
#document
| <td>

#data
<colgroup><td>
#errors
#document-fragment
tr
#document
| <td>

#data
<tbody><td>
#errors
#document-fragment
tr
#document
| <td>

#data
<tfoot><td>
#errors
#document-fragment
tr
#document
| <td>

#data
<thead><td>
#errors
#document-fragment
tr
#document
| <td>

#data
<tr><td>
#errors
#document-fragment
tr
#document
| <td>

#data
</table><td>
#errors
#document-fragment
tr
#document
| <td>

#data
<td><table></table><td>
#errors
#document-fragment
tr
#document
| <td>
|   <table>
| <td>

#data
<td><table></table><td>
#errors
#document-fragment
tr
#document
| <td>
|   <table>
| <td>

#data
<caption><a>
#errors
#document-fragment
td
#document
| <a>

#data
<col><a>
#errors
#document-fragment
td
#document
| <a>

#data
<colgroup><a>
#errors
#document-fragment
td
#document
| <a>

#data
<tbody><a>
#errors
#document-fragment
td
#document
| <a>

#data
<tfoot><a>
#errors
#document-fragment
td
#document
| <a>

#data
<th><a>
#errors
#document-fragment
td
#document
| <a>

#data
<thead><a>
#errors
#document-fragment
td
#document
| <a>

#data
<tr><a>
#errors
#document-fragment
td
#document
| <a>

#data
</table><a>
#errors
#document-fragment
td
#document
| <a>

#data
</tbody><a>
#errors
#document-fragment
td
#document
| <a>

#data
</td><a>
#errors
#document-fragment
td
#document
| <a>

#data
</tfoot><a>
#errors
#document-fragment
td
#document
| <a>

#data
</thead><a>
#errors
#document-fragment
td
#document
| <a>

#data
</th><a>
#errors
#document-fragment
td
#document
| <a>

#data
</tr><a>
#errors
#document-fragment
td
#document
| <a>

#data
<table><td><td>
#errors
#document-fragment
td
#document
| <table>
|   <tbody>
|     <tr>
|       <td>
|       <td>

#data
</select><option>
#errors
#document-fragment
select
#document
| <option>

#data
<input><option>
#errors
#document-fragment
select
#document
| <option>

#data
<keygen><option>
#errors
#document-fragment
select
#document
| <option>

#data
<textarea><option>
#errors
#document-fragment
select
#document
| <option>

#data
</html><!--abc-->
#errors
#document-fragment
html
#document
| <head>
| <body>
| <!-- abc -->

#data
</frameset><frame>
#errors
#document-fragment
frameset
#document
| <frame>

#data
#errors
#document-fragment
html
#document
| <head>
| <body>
